refactor: split oversize attend / attend-chat modules (#78)
PR #77 (ADR-129) landed several files past the project's
500-line review threshold and 800-line priority-split threshold.
This change carries out the splits the issue itemized.

- chip.rs (760) → chip/{render,registry,routing}.rs with chip/mod.rs
  re-exporting the public surface unchanged.
- app.rs (632) → app/{mod,keys}.rs; the iocraft component closure
  delegates each keyboard branch to a free function in keys.rs that
  takes a &str + cursor and returns the new state.
- cmd/run.rs (636) → cmd/run/{mod,governor,tick}.rs. The per-tick
  body, engagement-curve builder, and self-reload helper live in
  tick.rs and operate against a TickState bundle so they are
  reachable from tests without the full cmd_run shell.
- attend-instances/src/lib.rs (829) — left as-is per issue
  guidance; bulk is #[cfg(test)] code. Added a header note so the
  next reviewer doesn't re-flag.

No functional changes. New tests cover the seams (key handlers,
governor window cap, build_engagement reachability).

Closes #78
