+++ tools/attend-chat/src/app/mod.rs
@@ -10,45 +10,35 @@
 //!
 //! No sidebar, no focus filter, no threading render — those land in
 //! follow-up PRs on ADR-120.
+//!
+//! The keyboard handler bodies live in [`keys`] so each branch is a
+//! free function that can be unit-tested without standing up an
+//! iocraft `App`. The closure here is a thin dispatcher: read state,
+//! call the handler, write state back.
+
+mod keys;
 
 use agent_identity::TermCaps;
 use async_channel::Receiver;
 use iocraft::prelude::*;
 
-use crate::chip::{chip_for, color_for, known_identities, resolve_nickname, CHIP_WIDTH};
-use crate::sessions::discover as discover_sessions;
-use crate::text_layout::{render_cursor, split_at_char, visual_line_count};
-use crate::groups::{channels, live_peer_count, resolve_group_dir};
+use crate::chip::{chip_for, color_for, known_identities, CHIP_WIDTH};
+use crate::groups::channels;
 use crate::helper::{self, HelperMode};
-use crate::legend::{
-    all_completions, all_group_completions, apply_completion, find_trailing_mention,
-    group_legend_row, legend_row, parse_addressed, Addressed, Sigil,
-};
-use crate::signal::{cwd_dir, write_broadcast, write_signal, Signal};
+use crate::legend::{find_trailing_mention, group_legend_row, legend_row, Sigil};
+use crate::sessions::discover as discover_sessions;
+use crate::signal::Signal;
 use crate::slash;
+use crate::text_layout::{render_cursor, visual_line_count};
+
+use keys::{
+    handle_backspace, handle_char_insert, handle_delete, handle_end, handle_enter, handle_home,
+    handle_newline_insert, handle_tab, EnterAction, TabCycle,
+};
 
 #[derive(Default, Props)]
 pub struct AppProps {
     pub receiver: Option<Receiver<Signal>>,
-}
-
-/// Tab-completion cycle state.
-///
-/// First Tab on `@T<Tab>` inserts the first prefix-matching candidate
-/// and stores the candidate list + sigil here. Each subsequent Tab
-/// (with the buffer still ending in the previously-inserted token)
-/// advances the index, wrapping at the end. Any other keystroke
-/// implicitly resets the cycle: the Tab handler checks that the
-/// buffer's tail still matches `candidates[index]`, and if it
-/// doesn't, treats the press as a fresh completion start.
-///
-/// Same model applies to `#group<Tab>` cycling — the sigil
-/// discriminates which pool to draw candidates from.
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct TabCycle {
-    sigil: Sigil,
-    candidates: Vec<String>,
-    index: usize,
 }
 
 /// Upper bound on the in-memory message buffer. At typical chat rates
@@ -124,249 +114,44 @@ pub fn App(props: &AppProps, mut hooks: Hooks) -> impl Into<AnyElement<'static>>
                     // protocol; Alt-Enter is a cross-terminal
                     // fallback.
                     let v = input.read().clone();
-                    let (before, after) = split_at_char(&v, cursor.get());
-                    let next = format!("{}\n{}", before, after);
+                    let (next, new_cursor) = handle_newline_insert(&v, cursor.get());
                     input.set(next);
-                    cursor.set(cursor.get() + 1);
+                    cursor.set(new_cursor);
                 }
                 KeyCode::Enter => {
-                    let msg = input.read().trim_end().to_string();
-                    if !msg.is_empty() {
-                        // Slash-command interceptor — stops `/help`,
-                        // `/whois`, etc. from leaking onto the signal
-                        // bus as plain messages. Runs before the
-                        // `@`/`#`/broadcast dispatch because slash
-                        // syntax is start-of-input only and unambiguous.
-                        if let Some((cmd, args)) = slash::parse(&msg) {
-                            match slash::dispatch(cmd, args) {
-                                slash::SlashOutcome::Ok(s) => {
-                                    status.set(s);
-                                    input.set(String::new());
-                                    cursor.set(0);
-                                }
-                                slash::SlashOutcome::Err(s) => {
-                                    // Leave the buffer intact so the
-                                    // user can edit + retry without
-                                    // retyping.
-                                    status.set(s);
-                                }
-                            }
-                            return;
+                    let v = input.read().clone();
+                    let sigs = signals.read().clone();
+                    match handle_enter(&v, &sigs) {
+                        EnterAction::None => {}
+                        EnterAction::ClearWithStatus(s) => {
+                            status.set(s);
+                            input.set(String::new());
+                            cursor.set(0);
                         }
-                        let caps = TermCaps::detect();
-                        let seeds = discover_sessions();
-                        // Local instance cache for this Enter-handler
-                        // invocation. Distinct from the per-render
-                        // cache because key handlers are not on the
-                        // render path — they fire on demand and
-                        // resolve against the current registry state.
-                        let instance_cache =
-                            attend_instances::SnapshotCache::new();
-                        let agents = known_identities(
-                            &signals.read(),
-                            &seeds,
-                            caps,
-                            &instance_cache,
-                        );
-                        let result = match parse_addressed(&msg) {
-                            Some(Addressed::Agent(name)) => {
-                                match resolve_nickname(name, &agents) {
-                                    Some(target_cwd) => write_signal(&cwd_dir(&target_cwd), &msg)
-                                        .map(|n| format!("sent → @{name}: {n}")),
-                                    None => Err(std::io::Error::new(
-                                        std::io::ErrorKind::NotFound,
-                                        format!("@{name}: unknown nickname"),
-                                    )),
-                                }
-                            }
-                            Some(Addressed::Group(name)) => match resolve_group_dir(name) {
-                                // Empty-group rejection (ADR-129
-                                // follow-up). Mirrors PR #75's
-                                // `attend send --focus` discipline:
-                                // a `#groupname` send to a group
-                                // with zero live members would land
-                                // in `@<name>/` and sit unread until
-                                // cleanup, with the chat reporting
-                                // "sent" so the user assumes
-                                // delivery. The base channel
-                                // (`#open`) bypasses the check —
-                                // it rides `_broadcast/`, where
-                                // every attend scans regardless
-                                // of group membership.
-                                Some(dir) if live_peer_count(name) > 0 => {
-                                    write_signal(&dir, &msg)
-                                        .map(|n| format!("sent → #{name}: {n}"))
-                                }
-                                Some(_) => Err(std::io::Error::new(
-                                    std::io::ErrorKind::Other,
-                                    format!(
-                                        "#{name}: no live peers — message would not be read. \
-                                         Try #open to broadcast."
-                                    ),
-                                )),
-                                None => Err(std::io::Error::new(
-                                    std::io::ErrorKind::NotFound,
-                                    format!("#{name}: unknown group"),
-                                )),
-                            },
-                            None => write_broadcast(&msg).map(|n| format!("sent: {n}")),
-                        };
-                        match result {
-                            Ok(s) => {
-                                status.set(s);
-                                input.set(String::new());
-                                cursor.set(0);
-                            }
-                            Err(e) => status.set(format!("send failed: {}", e)),
-                        }
+                        EnterAction::StatusOnly(s) => status.set(s),
                     }
                 }
                 KeyCode::Tab => {
-                    // Autocomplete a trailing mention. First Tab on
-                    // `@partial` inserts the first matching candidate
-                    // and remembers the cycle (candidates +
-                    // index + sigil). Subsequent Tabs advance through
-                    // the cycle, wrapping at the end. A keystroke
-                    // that breaks the tail (typing a letter, moving
-                    // the cursor, etc.) silently invalidates the
-                    // cycle on the next Tab — the buffer's trailing
-                    // `@<name> ` no longer matches `candidates[index]`,
-                    // so we restart fresh from whatever partial is
-                    // currently typed.
-                    //
-                    // No-op if the caret isn't at the end of the
-                    // buffer — avoids surprising edits mid-sentence.
                     let buf = input.read().clone();
-                    let pos = cursor.get();
-                    if pos != buf.chars().count() {
-                        tab_cycle.set(None);
-                        return;
-                    }
-                    // Slash completion runs first — its grammar is
-                    // start-of-input-only, so when it matches, the
-                    // `@`/`#` logic below never applies. Silent
-                    // fall-through on no match so a stale `/` with
-                    // no registry hit doesn't trap the user.
-                    if let Some(partial) = slash::find_slash_partial(&buf) {
-                        if let Some(hit) = slash::best_slash_completion(partial.partial) {
-                            let (next, new_cursor) = slash::apply_slash_completion(hit.name);
-                            input.set(next);
-                            cursor.set(new_cursor);
-                        }
-                        // Slash and `@`/`#` share the cycle slot;
-                        // resetting prevents a stale agent cycle
-                        // from carrying across a slash insertion.
-                        tab_cycle.set(None);
-                        return;
-                    }
-                    let caps = TermCaps::detect();
-
-                    // Cycling: if we have a stored cycle and the
-                    // buffer still ends with the candidate we
-                    // just inserted (sigil + name + trailing
-                    // space, the exact form `apply_completion`
-                    // produces), advance to the next candidate.
-                    let cycle_now = tab_cycle.read().clone();
-                    if let Some(c) = cycle_now {
-                        let sigil_char = match c.sigil {
-                            Sigil::Agent => '@',
-                            Sigil::Group => '#',
-                        };
-                        let expected_tail =
-                            format!("{}{} ", sigil_char, c.candidates[c.index]);
-                        if buf.ends_with(&expected_tail) && !c.candidates.is_empty() {
-                            let next_idx =
-                                (c.index + 1) % c.candidates.len();
-                            let next = &c.candidates[next_idx];
-                            // Splice the old tail off and append
-                            // the new one. We work in chars rather
-                            // than bytes so multi-byte names stay
-                            // intact.
-                            let total_chars = buf.chars().count();
-                            let tail_chars =
-                                expected_tail.chars().count();
-                            let head: String = buf
-                                .chars()
-                                .take(total_chars - tail_chars)
-                                .collect();
-                            let new_buf =
-                                format!("{}{}{} ", head, sigil_char, next);
-                            let new_cursor = new_buf.chars().count();
-                            input.set(new_buf);
-                            cursor.set(new_cursor);
-                            tab_cycle.set(Some(TabCycle {
-                                index: next_idx,
-                                ..c
-                            }));
-                            return;
-                        }
-                    }
-
-                    // Fresh cycle: derive candidates from the current
-                    // trailing mention.
-                    let Some(mention) = find_trailing_mention(&buf) else {
-                        tab_cycle.set(None);
-                        return;
-                    };
-                    let candidates: Vec<String> = match mention.sigil {
-                        Sigil::Agent => {
-                            let seeds = discover_sessions();
-                            let instance_cache =
-                                attend_instances::SnapshotCache::new();
-                            let agents = known_identities(
-                                &signals.read(),
-                                &seeds,
-                                caps,
-                                &instance_cache,
-                            );
-                            all_completions(mention.partial, &agents)
-                        }
-                        Sigil::Group => {
-                            let groups = channels(caps);
-                            all_group_completions(mention.partial, &groups)
-                        }
-                    };
-                    if candidates.is_empty() {
-                        tab_cycle.set(None);
-                        return;
-                    }
-                    let first = candidates[0].clone();
-                    let (next_buf, new_cursor) =
-                        apply_completion(&buf, &mention, &first);
-                    input.set(next_buf);
-                    cursor.set(new_cursor);
-                    tab_cycle.set(Some(TabCycle {
-                        sigil: mention.sigil,
-                        candidates,
-                        index: 0,
-                    }));
+                    let cur = cursor.get();
+                    let cycle = tab_cycle.read().clone();
+                    let sigs = signals.read().clone();
+                    let res = handle_tab(&buf, cur, cycle, &sigs);
+                    input.set(res.new_buf);
+                    cursor.set(res.new_cursor);
+                    tab_cycle.set(res.new_cycle);
                 }
                 KeyCode::Backspace => {
                     let v = input.read().clone();
-                    let pos = cursor.get();
-                    if pos == 0 {
-                        return;
-                    }
-                    let (before, after) = split_at_char(&v, pos);
-                    // Drop one char off the end of `before`.
-                    let trimmed: String = before
-                        .chars()
-                        .take(pos.saturating_sub(1))
-                        .collect();
-                    input.set(format!("{}{}", trimmed, after));
-                    cursor.set(pos - 1);
+                    let (next, new_cursor) = handle_backspace(&v, cursor.get());
+                    input.set(next);
+                    cursor.set(new_cursor);
                 }
                 KeyCode::Delete => {
                     let v = input.read().clone();
-                    let pos = cursor.get();
-                    let total = v.chars().count();
-                    if pos >= total {
-                        return;
-                    }
-                    let (before, after) = split_at_char(&v, pos);
-                    // Drop the first char of `after`.
-                    let trimmed: String = after.chars().skip(1).collect();
-                    input.set(format!("{}{}", before, trimmed));
+                    let (next, new_cursor) = handle_delete(&v, cursor.get());
+                    input.set(next);
+                    cursor.set(new_cursor);
                 }
                 KeyCode::Left => {
                     let p = cursor.get();
@@ -382,38 +167,21 @@ pub fn App(props: &AppProps, mut hooks: Hooks) -> impl Into<AnyElement<'static>>
                     }
                 }
                 KeyCode::Home => {
-                    // Jump to the start of the current logical line.
                     let v = input.read().clone();
-                    let p = cursor.get();
-                    let (before, _) = split_at_char(&v, p);
-                    let line_start = before.rfind('\n').map(|i| {
-                        // chars-before count = char count of
-                        // `before[..=i]`.
-                        before[..=i].chars().count()
-                    });
-                    cursor.set(line_start.unwrap_or(0));
+                    cursor.set(handle_home(&v, cursor.get()));
                 }
                 KeyCode::End => {
-                    // Jump to the end of the current logical line.
                     let v = input.read().clone();
-                    let p = cursor.get();
-                    let (_, after) = split_at_char(&v, p);
-                    let line_end_in_after = after.find('\n');
-                    let add = match line_end_in_after {
-                        Some(byte) => after[..byte].chars().count(),
-                        None => after.chars().count(),
-                    };
-                    cursor.set(p + add);
+                    cursor.set(handle_end(&v, cursor.get()));
                 }
                 KeyCode::Char(c)
                     if !modifiers
                         .intersects(KeyModifiers::CONTROL | KeyModifiers::ALT) =>
                 {
                     let v = input.read().clone();
-                    let pos = cursor.get();
-                    let (before, after) = split_at_char(&v, pos);
-                    input.set(format!("{}{}{}", before, c, after));
-                    cursor.set(pos + 1);
+                    let (next, new_cursor) = handle_char_insert(&v, cursor.get(), c);
+                    input.set(next);
+                    cursor.set(new_cursor);
                 }
                 _ => {}
             },
@@ -628,5 +396,3 @@ pub fn App(props: &AppProps, mut hooks: Hooks) -> impl Into<AnyElement<'static>>
         }
     }
 }
-
-
